차이가 있을 때 블렌딩, 물 안에서 나무 생성 금지, float math.h 함수 사용

+++ perlin.c
@@ -43,7 +43,7 @@ static const float grad(const int p)
 
 const float perlin_noise(const float x)
 {
-    const int xi = (int)floor(x);
+    const int xi = (int)floorf(x);
     const float xf = x - xi;
 
     return lerp(grad(xi) * xf, grad(xi + 1) * (xf - 1.0f), fade(xf));
